Minor update to test/features/step_definitions/clockStepDefinitions.js
Turns out that the cwd of the script is set to the same dir as
clockStepDefinitions.js.  This, plus a change to the calling feature,
should allow html pages to be called off the filesystem.  This also
breaks calling of web pages with the same step, but that's easily fixed
if that bridge is crossed.